+++ mock-time/src/Control/Monad/Time.hs
@@ -9,11 +9,12 @@ where
 import Control.Monad.Trans
 import Data.Time hiding (getCurrentTime)
 
--- | A class interface with overridable time implementation.
--- Import `Control.Monad.Time.DefaultInstance` for the default
--- implementation using the real underlying time functions.
+-- | A class abstracting the implementation of time in Haskell.
+-- This module provides an overlappable default instance for monads
+-- implementing MonadIO.
+--
 -- See `Control.Monad.Trans.SimulatedTime` for an instance
--- that allows to control time.
+-- that allows to control time in tests.
 class Monad m => MonadTime m where
   getCurrentTime :: m UTCTime
   threadDelay :: Int -> m ()
@@ -27,3 +28,8 @@ instance {-# OVERLAPPABLE #-}
   where
   getCurrentTime = lift getCurrentTime
   threadDelay a = lift (threadDelay a)
+
+instance {-# OVERLAPPABLE #-}
+  (Monad m, MonadIO  m) => MonadTime m where
+  getCurrentTime = liftIO getCurrentTime
+  threadDelay a = liftIO (threadDelay a)