+++ mock-time/changelog.md
@@ -1,3 +1,8 @@
+## 0.2.0 (upcoming)
+  * Remove Control.Monad.Time.DefaultInstance in favour of an overlappable instance based on MonadIO in Control.Monad.Time
+  * Remove RealTimeT which can be replaced by simply using the instance from Control.Monad.Time
+  * Improve documentation
+
 ## 0.1.1 (2024-11-10)
 
   * Make it compile in ghc 9.6.6